extract widgets
